requirements.txt
defaultlist
django
pyahocorasick
regex